+++ src/planning.cpp
@@ -115,7 +115,7 @@ int main(int argc, char **argv) {
     while (ros::ok()) {
         ros::spinOnce();
 
-        float elapsed_time = ai_controller.observation.getTimeStamp();
+        double elapsed_time = ai_controller.observation.getTimeStamp();
         // printf("%f\n", elapsed_time);
         if(elapsed_time > 600 && elapsed_time != 0.0) {
           break;
